Run builds by platform to avoid creating multiple drafts
The matrix key's order defines which job runs first[0], so this
change will hopefully make all Windows builds ran before their
macOS counter-parts, effectively avoiding the race condition
between two `release.js` scripts listing the releases at the
same time - and then both creating a draft each.

It's not a hard guarantee, but if it may be a cheap workaround -
and it's not that much of an issue if the workaround fail.

[0] https://docs.github.com/en/actions/reference/workflow-syntax-for-github-actions#jobsjob_idstrategymatrix