Point production collectors to *.maap.aslm.org
We've migrated the maap-stores under the maap.aslm.org domain,
so this commit updates the collectors to use the new URLs.

Old collectors still work due to redirections set on the old
maap-xx.instedd.org domains to the new locations, but this
change will avoid redundant requests (collectors now make a
request to the instedd.org subdomain, get a redirection as
response, then repeate the request to the maap.aslm.org
subdomain).

We haven't migrated staging (at least yet), so we are not
upgrading that.